message

+++ week_9/session-1/pointer_arithmetic-1.c
@@ -2,15 +2,15 @@
 
 int main()
 {
-    int vector[5] = {11,13,17, 23 , 31};
+    float vector[5] = {11,13,17, 23 , 31};
 
-    int* address = vector;
+    float* address = vector;
 
-    printf ( "\nAddress[0] = %d\n", *address);
+    printf ( "\nAddress[0] = %f\n", *address);
 
-    printf ( "\nAddress[1] = %d\n", *(address+1));
+    printf ( "\nAddress[1] = %f\n", * (address+1) );
 
-    printf ("\n Address[2] = %d\n", *(address+2));
+    printf ("\n Address[2] = %f\n", *(address+2));
